Create llm-websearch.bash
First commit.  Just getting the bash script into github.  Adds the basic logic of searching SearxNG for about 25-30 URLs (if searx returns a full page of results) and has the LLM go through them one by one to decide if they are useful to answering the users ${question}.  If so the LLM summarizes the page into a file and moves to the next.